move t_seg_details, t_chan_seg_details, and t_chan_details to rr_types.h
